requirements.txt
streamlit
python-dotenv
google-generativeai
matplotlib
seaborn
urlextract
wordcloud
pandas
emojix